[proliferate] consolidate workspace naming mcp server
