+++ anyharness/crates/anyharness-lib/src/sessions/workspace_naming/mcp_server/mod.rs
@@ -1,0 +1,4 @@
+pub mod auth;
+mod protocol;
+pub mod server;
+mod tools;
