+++ anyharness/crates/anyharness-lib/src/sessions/workspace_naming/mcp_server/protocol.rs
