+++ anyharness/crates/anyharness-lib/src/sessions/workspace_naming/mod.rs
@@ -1,5 +1,3 @@
 pub mod eligibility;
 pub mod hooks;
-pub mod mcp;
-pub mod mcp_auth;
-mod mcp_protocol;
+pub mod mcp_server;
